moved rag feature to lib/tool
