+++ src/app/api/agents/documentQ&A/AIChat/query/route.ts
@@ -9,7 +9,7 @@ import {
     type CompanySearchOptions,
     type DocumentSearchOptions,
     type SearchResult
-} from "~/server/rag";
+} from "~/lib/tools/rag";
 import { validateRequestBody, QuestionSchema } from "~/lib/validation";
 import { auth } from "@clerk/nextjs/server";
 import { qaRequestCounter, qaRequestDuration } from "~/server/metrics/registry";
