+++ src/lib/tools/index.ts
@@ -1,0 +1,8 @@
+export {
+  runDocIngestionTool,
+  type DocIngestionToolInput,
+  type DocIngestionToolResult,
+  type DocIngestionToolRuntimeOptions,
+} from "./doc-ingestion";
+
+export * from "./rag";
