+++ src/lib/tools/rag/agentic/index.ts
@@ -1,0 +1,1 @@
+export { ragSearchTool, executeRAGSearch } from "./rag-search-tool";
